+++ src/main/java/roomescape/admin/reservation/AdminReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.reservation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,20 +6,20 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.ReservationService;
 
 @RestController
 @RequestMapping("/api/admin/reservations")
 public class AdminReservationController {
-    private final ReservationService reservationService;
 
-    public AdminReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
+    private final AdminReservationService adminReservationService;
+
+    public AdminReservationController(AdminReservationService adminReservationService) {
+        this.adminReservationService = adminReservationService;
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable long id) {
-        reservationService.deleteByAdmin(id);
+        adminReservationService.delete(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
