+++ src/test/java/roomescape/admin/AdminReservationControllerTest.java
@@ -9,9 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+import roomescape.admin.reservation.AdminReservationController;
+import roomescape.admin.reservation.AdminReservationService;
 import roomescape.exception.InvalidStateException;
 import roomescape.exception.NotFoundException;
-import roomescape.reservation.ReservationService;
 
 @WebMvcTest(AdminReservationController.class)
 class AdminReservationControllerTest {
@@ -20,7 +21,7 @@ class AdminReservationControllerTest {
     private MockMvc mockMvc;
 
     @MockitoBean
-    private ReservationService reservationService;
+    private AdminReservationService adminReservationService;
 
     @Test
     void 예약_삭제() throws Exception {
@@ -31,7 +32,7 @@ class AdminReservationControllerTest {
     @Test
     void 지난_날짜_예약_삭제시_400() throws Exception {
         willThrow(new InvalidStateException("이미 지난 날짜와 시간입니다."))
-                .given(reservationService).deleteByAdmin(1L);
+                .given(adminReservationService).delete(1L);
 
         mockMvc.perform(delete("/api/admin/reservations/1"))
                 .andExpect(status().isBadRequest());
@@ -40,7 +41,7 @@ class AdminReservationControllerTest {
     @Test
     void 존재하지_않는_예약_삭제시_404() throws Exception {
         willThrow(new NotFoundException("예약을 찾을 수 없습니다."))
-                .given(reservationService).deleteByAdmin(1L);
+                .given(adminReservationService).delete(1L);
 
         mockMvc.perform(delete("/api/admin/reservations/1"))
                 .andExpect(status().isNotFound());
