+++ src/test/java/roomescape/admin/AdminReservationTimeControllerTest.java
@@ -15,6 +15,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+import roomescape.admin.time.AdminReservationTimeController;
 import roomescape.exception.AlreadyInUseException;
 import roomescape.time.ReservationTimeService;
